Se agregaron validateData en el update y add de services y se arregraron alertas

+++ FrontEnd/src/components/AServices/AServiceTable.jsx
@@ -20,18 +20,40 @@ export const AServiceTable = () => {
 
   const deleteService = async (id) => {
     try {
-      const { data } = await axios.delete(`http://localhost:3100/services/delete/${id}`, {
+      const result = await Swal.fire({
+        title: 'you are sure?',
+        text: 'Delete Servicio aditional',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonText: 'YES',
+        cancelButtonText: 'CANCEL',
+      });
+        if (result.isConfirmed) {
+         
+          const { data } = await axios.delete(`http://localhost:3100/services/delete/${id}`, {
         headers: {
           'Authorization': token
         }
       })
+      if(data.message=== 'Service deleted'){ 
+        Swal.fire({
+            title: data.message ,
+              icon: 'success',
+              timer: 2000
+        })
+    }else{
       Swal.fire({
-        title: 'Deleted!',
-        text: 'Your file has been deleted.',
-        icon: 'success'
-      }).then(() => {
-        getServices()
+          title: data.message ,
+            icon: 'warning',
+            timer: 2000
       })
+  }
+    
+     getServices()
+        } 
+        //
+      
+      
     } catch (err) {
       console.error(err)
       alert(err.response.data.message)
